requirements.txt
opencv-python==4.11.0.86
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.2.0
tqdm==4.66.1
streamlit==1.36.0 